chore: move FunctionDecl in Decl directory
