+++ test/AST/Decl/FunctionDecl.cpp
@@ -1,4 +1,4 @@
-#include "FunctionDecl.hpp"
+#include "Decl/FunctionDecl.hpp"
 
 #include <gtest/gtest.h>
