+++ test/CMakeLists.txt
@@ -27,7 +27,7 @@ target_sources (unit_tests
     PRIVATE
     main.cpp
     AST/ASTNode.cpp
-    AST/FunctionDecl.cpp
+    AST/Decl/FunctionDecl.cpp
 )
 
 include (GoogleTest)
